+++ packages/api-derive/src/staking/cache.ts
@@ -26,13 +26,13 @@ export function getEraMultiCache <T> (CACHE_KEY: string, eras: EraIndex[], withA
   return cached;
 }
 
-export function setEraCache <T extends { era: EraIndex }> (cacheKey: string, withActive: boolean, value: T): T {
+export function setEraCache <T extends { era: EraIndex }> (cacheKey: string, value: T, withActive?: boolean): T {
   !withActive && deriveCache.set(cacheKey, value);
 
   return value;
 }
 
-export function setEraMultiCache <T extends { era: EraIndex }> (CACHE_KEY: string, withActive: boolean, values: T[]): T[] {
+export function setEraMultiCache <T extends { era: EraIndex }> (CACHE_KEY: string, values: T[], withActive?: boolean): T[] {
   !withActive && values.forEach((v) => deriveCache.set(`${CACHE_KEY}-${v.era.toString()}`, v));
 
   return values;