+++ packages/api-derive/src/staking/erasPoints.ts
@@ -34,8 +34,8 @@ function mapPoints (eras: EraIndex[], points: PalletStakingEraRewardPoints[]): D
   }));
 }
 
-export function _erasPoints (instanceId: string, api: DeriveApi): (eras: EraIndex[], withActive: boolean) => Observable<DeriveEraPoints[]> {
-  return memo(instanceId, (eras: EraIndex[], withActive: boolean): Observable<DeriveEraPoints[]> => {
+export function _erasPoints (instanceId: string, api: DeriveApi): (eras: EraIndex[], withActive?: boolean) => Observable<DeriveEraPoints[]> {
+  return memo(instanceId, (eras: EraIndex[], withActive?: boolean): Observable<DeriveEraPoints[]> => {
     if (!eras.length) {
       return of([]);
     }
@@ -46,7 +46,7 @@ export function _erasPoints (instanceId: string, api: DeriveApi): (eras: EraInde
     return !remaining.length
       ? of(cached)
       : api.query.staking.erasRewardPoints.multi(remaining).pipe(
-        map((p) => filterCachedEras(eras, cached, setEraMultiCache(CACHE_KEY, withActive, mapPoints(remaining, p))))
+        map((p) => filterCachedEras(eras, cached, setEraMultiCache(CACHE_KEY, mapPoints(remaining, p), withActive)))
       );
   });
 }