+++ packages/api-derive/src/staking/ownExposure.ts
@@ -10,8 +10,8 @@ import { combineLatest, map, of } from 'rxjs';
 import { firstMemo, memo } from '../util/index.js';
 import { erasHistoricApplyAccount } from './util.js';
 
-export function _ownExposures (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive: boolean) => Observable<DeriveOwnExposure[]> {
-  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive: boolean): Observable<DeriveOwnExposure[]> =>
+export function _ownExposures (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive?: boolean) => Observable<DeriveOwnExposure[]> {
+  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive?: boolean): Observable<DeriveOwnExposure[]> =>
     eras.length
       ? combineLatest([
         combineLatest(eras.map((e) => api.query.staking.erasStakersClipped(e, accountId))),