+++ packages/api-derive/src/staking/stakerPrefs.ts
@@ -10,8 +10,8 @@ import { map } from 'rxjs';
 import { memo } from '../util/index.js';
 import { erasHistoricApplyAccount } from './util.js';
 
-export function _stakerPrefs (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive: boolean) => Observable<DeriveStakerPrefs[]> {
-  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive: boolean): Observable<DeriveStakerPrefs[]> =>
+export function _stakerPrefs (instanceId: string, api: DeriveApi): (accountId: Uint8Array | string, eras: EraIndex[], withActive?: boolean) => Observable<DeriveStakerPrefs[]> {
+  return memo(instanceId, (accountId: Uint8Array | string, eras: EraIndex[], _withActive?: boolean): Observable<DeriveStakerPrefs[]> =>
     api.query.staking.erasValidatorPrefs.multi(eras.map((e) => [e, accountId])).pipe(
       map((all): DeriveStakerPrefs[] =>
         all.map((validatorPrefs, index): DeriveStakerPrefs => ({